Use `System.IO.Path.Combine` any time there is a path, except when we can't
In the cases where we can't, use the correct path separator for the platform.
All of said cases are split by platform. If we end up with cases where they aren't, it would make a lot of sense to use `/` as the path separator as even Windows supports this.

This commit also makes the indentation type consistent in the relevant SteamAudio.NET files.